pyproject.toml
[project]
name = "imagoimages"
version = "0.1.0"
description = ""
authors = [
    {name = "Shashank Kumar",email = "shashankkumarkushwaha@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "elasticsearch-dsl (>=8.17.1,<9.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-filter (>=25.1,<26.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.test.dependencies]
poetry-plugin-shell = "^1.0.1"
pytest-django = "^4.10.0"
